add docker compatibility

+++ src/main/java/hk/zdl/crypto/peth/plot/gui/Main.java
@@ -259,7 +259,11 @@ public class Main {
 	private static void do_plot(Path dir, String id, long nounce, PlotProgressListener listener) throws Exception {
 		Path plotter_bin_path = copy_plotter().toPath();
 		Process proc = PlotUtil.plot(plotter_bin_path, dir, false, new BigInteger(id), Math.abs(new Random().nextInt()), nounce, listener);
-		proc.waitFor();
+		int i = proc.waitFor();
+		if (i != 0) {
+			var err_info = IOUtils.readLines(proc.getErrorStream(),Charset.defaultCharset()).stream().reduce("",(a,b)->a+"\n"+b).trim();
+			throw new IOException(err_info);
+		}
 		Files.deleteIfExists(plotter_bin_path);
 	}
